Initial Commit
A short Project to comply with an Interview for Uken Games.
The main purpose of the project is to find the least frequent number in a given dataset of n terms from a formatted text file. If there is a collision, the lowest number sharing the same frequency is reported instead. 
In-depth Instructions are in the Description.